remove my completionSlice and move the functions into existing slice file
