+++ store/taskCompletionSlice.ts
@@ -1,14 +1,16 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { Task, TaskCompletion } from "../types";
 import { useAppSelector } from "./store";
-import { taskCompletions } from "../data";
+import { profiles, taskCompletions } from "../data";
 
 interface TaskCompletionState {
   taskCompletions: TaskCompletion[];
+  avatarInCompletion:string;
 }
 
 const initialState: TaskCompletionState = {
   taskCompletions: taskCompletions,
+  avatarInCompletion:"",
 };
 
 const taskCompletionSlice = createSlice({
@@ -34,9 +36,37 @@ const taskCompletionSlice = createSlice({
 
       state.taskCompletions.push(newTaskCompletion);
     },
+    findCompletionsByTaskId: (state, action: PayloadAction<{ taskId: string }>) => {
+        const { taskId: id } = action.payload;
+        const foundCompletion = state.taskCompletions.find((completion) => completion.id === id);
+  
+        if (foundCompletion) {
+          state.taskCompletions = [foundCompletion]; 
+        } else {
+        }
+      },
+      findCompletionsByTaskAndProfielId: (state, action: PayloadAction<{ taskId: string, profielId:string }>) => {
+        const { taskId,profielId } = action.payload;
+        const foundCompletion = state.taskCompletions.find((completion) => completion.taskId=== taskId && completion.profileId === profielId);
+        if (foundCompletion) {
+          state.taskCompletions = [foundCompletion]; 
+        } else {
+        }
+      },
+      findAvatarInCompletionByTaskId: (state, action: PayloadAction<{ taskId: string }>) => {
+        const { taskId } = action.payload;
+        const foundCompletion = state.taskCompletions.find((completion) => completion.taskId === taskId);
+        const profile = profiles.find ((p)=>p.id === foundCompletion?.profileId);
+
+        if (profile) {
+          const foundAvatar = profile.avatar
+          state.avatarInCompletion = foundAvatar;
+
+        }
+      },
   },
 });
 
-export const { setTaskAsCompleted } = taskCompletionSlice.actions;
+export const { setTaskAsCompleted,findCompletionsByTaskId,findCompletionsByTaskAndProfielId,findAvatarInCompletionByTaskId } = taskCompletionSlice.actions;
 
 export const taskCompletionReducer = taskCompletionSlice.reducer;
